Extract write to file helper func
