+++ pkg/sdk/poc/main.go
@@ -10,6 +10,7 @@ import (
 	"os"
 	"strings"
 
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/internal/gencommons"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk/poc/example"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk/poc/generator"
@@ -103,7 +104,7 @@ func runAllTemplatesAndSave(definition *generator.Interface, file string) {
 func runTemplateAndSave(def *generator.Interface, genFunc func(io.Writer, *generator.Interface), fileName string) {
 	buffer := bytes.Buffer{}
 	genFunc(&buffer, def)
-	generator.WriteCodeToFile(&buffer, fileName)
+	gencommons.WriteCodeToFile(&buffer, fileName)
 }
 
 func filenameFor(prefix string, part string) string {
